+++ components/script/dom/webidls/ReadableStreamDefaultReader.webidl
@@ -32,3 +32,16 @@ dictionary ReadableStreamReadResult {
  boolean done;
 };
 
+// The TeeReadRequest interface is entirely internal to Servo, and should not be accessible to
+// web pages.
+[LegacyNoInterfaceObject, Exposed=(Window,Worker)]
+// Need to escape "TeeReadRequest" so it's treated as an identifier.
+interface _TeeReadRequest {
+};
+
+// The TeeUnderlyingSource interface is entirely internal to Servo, and should not be accessible to
+// web pages.
+[LegacyNoInterfaceObject, Exposed=(Window,Worker)]
+// Need to escape "TeeUnderlyingSource" so it's treated as an identifier.
+interface _TeeUnderlyingSource {
+};
