+++ components/script/microtask.rs
@@ -22,7 +22,7 @@ use crate::dom::globalscope::GlobalScope;
 use crate::dom::htmlimageelement::ImageElementMicrotask;
 use crate::dom::htmlmediaelement::MediaElementMicrotask;
 use crate::dom::mutationobserver::MutationObserver;
-use crate::dom::readablestreamdefaultreader::TeeReadRequestMicrotask;
+use crate::dom::teereadrequest::TeeReadRequestMicrotask;
 use crate::realms::enter_realm;
 use crate::script_runtime::{notify_about_rejected_promises, CanGc, JSContext};
 use crate::script_thread::ScriptThread;
